+++ src/integTest/java/com/google/cloud/tools/gradle/appengine/AppEngineStandardPluginIntegrationTest.java
@@ -27,6 +27,8 @@ import org.gradle.testkit.runner.GradleRunner;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -37,25 +39,62 @@ import java.io.IOException;
 import java.util.Arrays;
 
 /**
- * End to end tests for flexible environment projects
+ * End to end tests for standard projects
  */
-public class AppEngineFlexiblePluginIntegTest {
+public class AppEngineStandardPluginIntegrationTest {
+
+  @BeforeClass
+  public static void assertEnvironment() {
+    Assert.assertThat(System.getProperty("java.version"), CoreMatchers.startsWith("1.7"));
+  }
 
   @Rule
-  public Timeout globalTimeout = Timeout.seconds(600);
+  public Timeout globalTimeout = Timeout.seconds(60);
 
   @Rule
   public final TemporaryFolder testProjectDir = new TemporaryFolder();
 
   @Before
   public void setUp() throws IOException {
-    FileUtils.copyDirectory(new File("src/integTest/resources/projects/flexible-project"),
+    FileUtils.copyDirectory(new File("src/integTest/resources/projects/standard-project"),
         testProjectDir.getRoot());
+  }
+
+  @Ignore
+  @Test
+  public void testDevAppServer_sync() {
+    // TODO : write test for devapp server running in synchronous mode
+  }
+
+  /**
+   * If this test is failing, make sure you've set JAVA_HOME=some-jdk7, it might have something
+   * to do with the way dev_appserver.py is launching java.
+   */
+  @Test
+  public void testDevAppServer_async() throws InterruptedException, IOException {
+    GradleRunner.create()
+        .withProjectDir(testProjectDir.getRoot())
+        .withPluginClasspath()
+        .withArguments("gcpAppStart")
+        .build();
+
+    AssertConnection.assertResponse("http://localhost:8080", 200,
+        "Hello from the App Engine Standard project.");
+
+    GradleRunner.create()
+        .withProjectDir(testProjectDir.getRoot())
+        .withPluginClasspath()
+        .withArguments("gcpAppStop")
+        .build();
+
+    // give the server a couple seconds to come down
+    Thread.sleep(8000);
+
+    AssertConnection.assertUnreachable("http://localhost:8080");
   }
 
   @Test
   public void testDeploy() throws ProcessRunnerException {
-
     BuildResult buildResult = GradleRunner.create()
         .withProjectDir(testProjectDir.getRoot())
         .withPluginClasspath()
@@ -64,11 +103,11 @@ public class AppEngineFlexiblePluginIntegTest {
         .build();
 
     Assert.assertThat(buildResult.getOutput(),
-        CoreMatchers.containsString("Deployed service [flexible-project]"));
+        CoreMatchers.containsString("Deployed service [standard-project]"));
 
     CloudSdk cloudSdk = new CloudSdk.Builder()
         .exitListener(new NonZeroExceptionExitListener())
         .build();
-    cloudSdk.runAppCommand(Arrays.asList("services", "delete", "flexible-project"));
+    cloudSdk.runAppCommand(Arrays.asList("services", "delete", "standard-project"));
   }
 }
